#1010 [REFACTOR] DimModal -> DimModalLayout 컴포넌트명 변경
