+++ src/shared/component/newModal/DimModalLayout/DimModalLayout.module.css
